requirements.txt
fastapi
uvicorn
slack_sdk
python-dotenv
langchain
chromadb
google-generativeai
pypdf
